remove selector printable, list discreteOwner
The selectors printable code wasn't being used in any
valid way, so that's out.  Backup List printing now refers
to the DiscreteOwner in the embedded selector as the
ResourceOwner reference.  Renamed the "Selectors"
column in the List print table to "Resource Owner" to
better match what should be contained in that field.