full backup if prev snapshots have no details
In the event that a backup is completed but the details,
somehow, isn't persisted, we want the next backup to
do a full, instead of an incremental, backup.  If we don't
have this protection the following backups could end up
in a bad state.  Future changes will add better resilience
so that the fallback isn't needed.