Dedupe items by ID in OneDrive Collection
It's possible for the same item to be returned multiple times when
enumerating items. Handle this by deduping by ID within a Collection.

This does not handle the case where the item may have been deleted from
the Collection.

This does not handle the case where the item may appear multiple times
but in different Collections (due to single endpoint).