Return file paths and delta URLs for OneDrive
Track and return the delta URL and file paths for selected items. File
path collection is a little odd right now when the selector is not
specifying everything. This is because we actually check the parent path
to see if we should include an item, even if that means we don't save
the folder containing the selected item to the path map.