Fail operation if we failed to fetch item data
If GraphConnector reported errors while trying to fetch item data, fail
the entire operation. This stops silent failures, but is a big hammer at
the moment because there's no checks for items no longer being
available.

Error reporting is minimal as well, but some info should be in the log.