Check for and ignore item not found errors
Item not found errors may appear because an item was moved or deleted
between enumerating the items in a folder and fetching the data for the
items in the folder. This keeps them from being reported as there is
nothing we can do if the item data is no longer available.