Populate and propagate del list for deleted files
Begin generating and propagting a list of items to exclude from the
upcoming backup when merging hierachies in kopia. Items in this list
should be globally unique.

It is not safe to return this list of excluded items until OneDrive uses
item IDs for kopia file names. Otherwise we can't guarantee globally
unique entries in the list and may accidentally exclude things we
shouldn't.