Expand unit test to ensure file not in kopia
Test that files that have errors reading data are not available for
restores. This protects us against kopia-assisted incrementals kicking
in on files that failed to upload properly but are thought to be cached.
If these files were restorable (and in the kopia snapshot), they could
hide real file changes because only the modtime for streaming files is
checked, not the file size.