Expand error checks for kopia backup operation
Kopia was not originally intended to be used as an SDK so the error
passing is a little thin in places. Snapshot operations may not return
an error object if errors were ignored.

Handle this by expanding the checks done in BackupOp to return an error
if kopia reported any number of errors. This is very conservative, but
will keep from marking snapshots as succeeded if they had errors with
anything.