Patch edge case for metadata deserialization
Currently no folder path entry is stored for the root folder. This leads
to not having a folders map but having a valid delta token if all items
are in the root of the drive.

Update the code to add at least an empty folders map entry so we don't
think we have missing metadata. Add tests to check for these edge cases.

Long-term we may want to add a folder entry for the root folder. Whether
we do or not may depend on how we decide to set the state field for that
collection.