Make calendars a 2-level hierarchy
All calendars except the default are nested under the default calendar.
Having a non-default calendar named the same as the default calendar
does not cause problems when fetching the default calendar by name. Only
the default calendar will be returned in that situation.

This fixes the bug where we had multiple collections for the same path
but representing different folders.