Create and return an interface instead of structs
Don't export the structs we wrap suite.Suite in. Instead return an
interface that (hopefully) has the functions we need. This makes things
less error prone as users of the wrapper are forced to call one of the
constructor functions, giving them no way to avoid the builtin checks
they provide.