Add Fetch function to RestoreCollection
This function allows for more of a random access to items in a
collection and can even allow getting items that weren't originally
going to be returned in the Items() channel. This function is different
than Items() because it's synchronous and meant to be called inline with
other code.

Broadly speaking, this pattern will allow us to remove ordering
dependencies between items during restore. For example, we won't have to
order metadatafiles in OneDrive after data files.