Update to PR template
This is a small QOL improvement to the PR template to fix some
annoyances.

When creating small PR(usually a few commits), the first commit is
usually has the title and description we need. GitHub does take this
information and dump it into the editor, but we end up having to cut
paste `## Description` from the end of what GitHub added to the
top. This change avoids that. Also, now that description is no more in
bold with underlines, I've reduced the heading level of the other
ones.