find calendars by id when restoring
Now that calenders are stored by ID, we have a
problem with restore destinations: the dest is
a name, not an ID.  Even though we create the
dest during restoration, we don't have a way to
reference the ID of the created dest, which
causes additional restores to fail on a "folder
already exists" post error.  This change adds in
the missing dest -> id lookup.