Fix OneDrive permissions not getting backed up
When we do a new backup, we look up the old permissions assuming they
are accurate, but the permissions from the previous backup might be
incorrect as we may have run it with permissions backup disabled.

It can also happen the other way round where even if we ask it to
disable permissions backup, it can still end up having permissions
which are pulled from the previous backup.

This fixes this behavior by setting the mod time for the backup
info to the current time.