Handle parentReference for deleted items not having path
parentReference for deleted items will not have path. This change
makes it so that we use the path from the previous backup to handle
those scenarios. Since we need the path of the collection from
previous backup, we now also store the path for the root collection
for lookup in later backups.