Refactor code to delete existing files last
Attempt to preserve as much data as possible in the face of errors by
adding updated/moved/renamed items to the exclude list last. This
ensures that even if we have errors in earlier parts of the code (i.e.
looking up the collection or previous collection) and we're running in
best-effort mode the previous version of the file will continue to
exist.