Update to PR template (#2458)
## Description

This is a small QOL improvement to the PR template to fix some annoyances.

When creating small PR(usually a few commits), the first commit is usually has the title and description we need. GitHub does take this information and dump it into the editor, but we end up having to cut paste `## Description` from the end of what GitHub added to the top. This change avoids that. Also, now that description is no more in bold with underlines, I've reduced the heading level of the other ones.

*Just a personal opinion, let me know if folks have a difference in opinion*.

## Does this PR need a docs update or release note?

- [ ] :white_check_mark: Yes, it's included
- [ ] :clock1: Yes, but in a later PR
- [x] :no_entry: No 

## Type of change

<!--- Please check the type of change your PR introduces: --->
- [ ] :sunflower: Feature
- [ ] :bug: Bugfix
- [ ] :world_map: Documentation
- [ ] :robot: Test
- [ ] :computer: CI/Deployment
- [ ] :broom: Tech Debt/Cleanup