Reduce total number of OneDrive tests run
Consolidate tests for OneDrive permissions so that we don't end up
running as many tests total. Behavior checked in old version of tests is
still checked in the consolidated test.

This consolidation will help keep data production under control and thus
reduce the chance of throttling in CI. The main driver behind this is
the addition of the ability to easily test different versions and the
desire to ensure old versions still behave as expected. However, with
the increasing number of versions we have this leads to many tests.