Reduce total number of OneDrive tests run (#2637)
Consolidate tests for OneDrive permissions so that we don't end up running as many tests total. Behavior checked in old version of tests is still checked in the consolidated test.

This consolidation will help keep data production under control and thus reduce the chance of throttling in CI. The main driver behind this is the addition of the ability to easily test different versions and the desire to ensure old versions still behave as expected. However, with the increasing number of versions we have this leads to many tests.

---

#### Does this PR need a docs update or release note?

- [ ] :white_check_mark: Yes, it's included
- [ ] :clock1: Yes, but in a later PR
- [x] :no_entry: No

#### Type of change

- [ ] :sunflower: Feature
- [ ] :bug: Bugfix
- [ ] :world_map: Documentation
- [x] :robot: Test
- [ ] :computer: CI/Deployment
- [x] :broom: Tech Debt/Cleanup

#### Issue(s)

* closes #2636

#### Test Plan

- [ ] :muscle: Manual
- [x] :zap: Unit test
- [ ] :green_heart: E2E