Augment paths for OneDrive restore
This adds the missing .dirmeta files which are needed to make sure
that permissions are being restored properly. Although the newer
version can fetch permissions file using the Fetch API, we still have
to populate the items as it cannot fetch items from previous
directories which is needed to make sure we can restore permissions
correctly. As of now, by adding a dirmeta entry for each folder that
is part of the items, we make sure that all folders get their own
collection which in turn will help us properly restore permissions for
all folders.