Fixup ShortRefs for OneDrive and SharePoint files
Have files in OneDrive and SharePoint take into account the file name
when generating the ShortRef. This ensures the ShortRef changes when the
file name does.

This change will not affect Corso's ability to restore previous backups
as each backup is independent of the others. This change can be reverted
eventually when SDK consumers do not require distinct ShortRefs when a
file name changes.