selectors use mulit-value short-refs
With onedrive storage file names being changed from
the file display name to the file id, we need a more
granular form of indentification when using selectors
to choose which values count as matchable fields.
This change modifies the selector PathValues to
return slices of strings for each category instead of
a single string, and the reducer matches on any.
This will allow each service to decide what values
are considered equivalent (id, shortRef, a value inside
the info, etc) for each property.