ShortRef munging for OneDrive/SharePoint (#2750)
Munge ShortRefs for OneDrive and SharePoint Libraries to
include the file name. This preserves previous behavior
where renaming a file would generate a new ShortRef for
said item, even if the file was not moved across
directories. This functionality can be removed at a
later time when SDK consumers have had a chance to
update their behavior

Plan is to merge this PR into #1535 and then merge into
main when that's ready so this all goes in at once

Manually tested renaming a file in the same folder and
verified backups had different ShortRefs for the
OneDrive file

---

#### Does this PR need a docs update or release note?

- [ ] :white_check_mark: Yes, it's included
- [ ] :clock1: Yes, but in a later PR
- [x] :no_entry: No

#### Type of change

- [x] :sunflower: Feature
- [ ] :bug: Bugfix
- [ ] :world_map: Documentation
- [ ] :robot: Test
- [ ] :computer: CI/Deployment
- [ ] :broom: Tech Debt/Cleanup

#### Issue(s)

* #1535

#### Test Plan

- [x] :muscle: Manual
- [x] :zap: Unit test
- [ ] :green_heart: E2E